Adds the continuation of an older project here.
The idea was to create a game by using design patterns and way to optimize the project. This was for two earlier courses I had in my education. I decided to continue working on it by and add AI related things such as steering behavior and formation pattern since AI in games is what I want to learn more about.