Fixes done to enemyBoar, overall changes in enemies, enemyManager and projectileManager. Player is now a part of the objectBase quadtree.
Uses a bool that decides if the player is in range for enemies to minimize the amount of IsInDistance if checks.
Fixed a bug with the boars prioritySteering which made it use only the spearation behavior if needed while ignoring all the other behaviors.
Damage, range and health values changed for enemies.
Damage and speed changed on projectiles.
EnemyManager keeps track on how many times it has spawned enemies in the survival state and adds that amount to the next wave of enemies.
Some value changes in the players health regen and projectile speed to make it stronger.
Now the game over state calls a new SurvivalGameState instead of the InGameState. Which is how it's supposed to work right now.
Using the quadtree in separationBehavior to see if the enemy collides with the player.
The player is the first object inserted in the quadtree, which makes it possible to use if(_objectsHit[0]->GetObjectID() == playerCharacter->GetObjectID())  when I only want to check for the player. If the first object isn't the player, then it hasn't collided with the player.
Also, changed the projectileManager, check collision to use a switch statement instead of if statement when it comes to the projectile types.