First attempt on making formations work as intended
Still work in progress, but pushes these changes for now since it works. Even though I will continues improving on these things.
Adds a new weapon component called shield, which gives the wielder more hp, but currently has no attack.
Add the use of an enum class that keeps tracks on the steering behaviors type.
Improves the VShape formation which still is a work in progress. Now it creates three rows of VShape formations that is using slot rolls to keep the shield users in front and staff users in the back.
Now the weapon components has an owner, so the weapon can get access to and change values for the enemy that wields it. Currently only used in the sword component to change the enemies target position to the players position.