Human updates its target, replace behavior function and adds an IsOutOfDistance method
Now the human enemy updates its target based on an enum class. Currently it is only useful if the human has the sword component, it checks if the player is within a certain range and changes target between player position and slot position. It also replaces the angular steering behavior based on the current target enum.
Now priority steering and blend steering has a function that locates a specific behavior based the behavior type and replaces it with a new one. If the specific behavior isn't found, it returns false. Also adds a function that simply removes the specific behavior.
Adds a new function that checks if the distance between to vectors is larger than the set distance.