Object pool for weapons and changed parameter in steering behavior.
Now there is an object pool for weapons which the enemyHuman gets access to from enemyManager instead of creating a new weapon everytime the enemy gets activated.
Changed the parameters in the steering behavior, now it takes in const ObjectBase to get access to position and object ID while also taking in the behaviorData struct to get access to the steering behavior values. This makes it possible for other objects aside from enemyBase to use the steeringBehaviors.