Finished with the formation state for now
Now the enemyManager spawns formations of enemies based on the current wave. For each wave, the number of enemies will increase.

The formationManager pickes between 2 formations: defensive circle and VShape.
Every slot has a role which is set at the creation of the slot. When assigning enemies to the slots, it will compare the weaponType with the role and pick the best suited slot. In DefensiveCircle, the outer circles has defender roles, middle has swordfighters and inner has mages. When spawning enemies, they will select weaponType according to the amount of slots for each role there is.
I changed the UpdateSlots function a little, now it uses and updates an anchorPoint which the formation follows and there is a check that makes sure all slots are inside the border and at the position. Once the slots are at position, the swordfighters will run to the player and attack.

Now the player shoots 3 projectiles at the same time. Mostly because I think multi shot is cool.