Cleaned up code and fixed typos
Now weaponeComponent is not a pure virtual and contains getters that all weapons share instead of repeating them.

Projectiles uses less parameters in the creation and when activated. Spritepaths for projectiles are saved in an unordered map and are accessed through the projectileType enum.
Now there is a third projectile that bosses will use.

Circle colliders for projectile and enemies now sets their radius based on the sprite size. Will only work with squared sprites, which works for now

Fixes a typo in the debugDrawer for circles.

Renamed the objectID variable to lastObjectID, to describe better what it is.

Adds a pointBetweenVector function which creates a point at a set distance from VectorA to VectorB.

Fixed typo in DecisionTreeAction.

Adds a function in Timer that sets the current time.
Fixes typo in activate timer.

Adds a sketch of a decision tree created in draw.io, which is the decision pattern the first boss will follow.