Some more cleaning up and adds things that will be used in the future.
Makes ObjectBase handle all the getter and setters for those that inherits from it, since it's not a pure virtual anymore.
Changes so the EnemyBase class handles the damage taken for all enemies that inherits from it.
Now ObjectBase has a currentTarget variable.
Now there is a TuskComponent which is a WeaponComponent that has a dash as attack.
Player uses a new WeaponComponent called SuperStaff which shoots the projectiles instead of the player.
Now there is a method called ClosestPosition which will be used to locate the closest position in a vector of positions.
Some minor changes done to the DecisionTree in the making.
Now there is a vector of playerCharacters. It's only one player currently, but the idea is to create more players in  the future. Similar changes will be made throughout the project.
Now projectiles keeps track of the ObjectBase that fires it. This makes it easier to keep track of what objectType it should damage.
Now the collider also keeps track of the ObjectBase it's connected to.