Optimization and cleaning up in dataStructuresAndMethods header and cpp
Removes temporary variables and creates them in the headers instead. It makes the compiling time a little longer, but the runtime a little smoother.
Now the dataStructuresAndMethods is called universalFunctions and is a class that contains functions used throughout the projects.
QuickSort and BinarySearch is moved to a new class called searchSortAlgorithms, it's still a work in progress on how I want to do on that part.
Creates a new class called managerBase which enemyManager and projectileManager inherits from. Since they shared some similarities, those got moved to the managerBase class.