Adds weaponPool and changes activeObjects to an unordered map instead of vector
Now weapon pool are handled by a weaponManager instead of the enemyManager. The weaponManager updates and renders all active weapons.
Changed the activeObjects vector to an unordered map with objectID as key. Now I don't have to use quickSort and binarySearch when removing specific enemies.
Now colliders got an isActive bool, if it's false, it won't get added to the quadTree.
Now formationsManager is called formationHandler, since it's not supposed to inherit from managerBase, this makes it less confusing.