Function for setting weapon values and continue working on decisionTree
Now weaponComponent has a SetValues funcion which is called by the playerCharacter and bossBoar since they require changes in the weapon values.
It also has a SetValuesToDefault finction which is used for the enemyHumans when it access a weapon.
Continued the work on decisionTree, now when I've done the changes I wanted to do in the project, so far.
Now there is 3 different attack actions nodes which is a dash, a stomp and a fire projectile action.
These will be used by the bossBoar.
The decisionTree in draw.io was also updated.