New resolution, formation mages has super staff and projectiles has a quadtree
Now the project starts with 1920 x 1080 resolution.
Mages has super staff instead of ordinary staff in the formation mode.
Projectiles are put in their own quadtree, this quadtree is querried by the other objects, aside from projectiles. This makes so projectiles doesn't have to go through all the other projectiles if they stack up. They are not going to collide with each other, so there is no need for them to loop through other projectiles.